debian/extra/ifup@.service: Additionally run for "auto" class
We don't really support "allow-hotplug" in Ubuntu at the moment, the ifupdown
scripts for bridge-utils, ifenslave, vlan, and others only look for "auto".

So we need to deal with "auto" devices appearing after "/etc/init.d/networking
start" already ran.

Also, check if devices are actually defined in /etc/network/interfaces as we
don't use Debian's net.agent.

Run ifup in the background during boot, to avoid blocking network.target on
boot with unavailable interfaces in /etc/network/interfaces. The main purpose
of the Before=network.target is to ensure a robust shutdown order, not to block
the boot on ifup'ing every interface; services which want that need to depend
on network-online.target.

LP: #1374521
LP: #1425376