73-usb-net-by-mac.rules: Test for disabling 80-net-setup-link.rules more efficiently
Do not run readlink for *every* uevent, this should (at most) be done for USB
network devices, so move that test after that.

Also, stop calling readlink at all and just test if
/etc/udev/rules.d/80-net-setup-link.rules exists -- a common way to disable an
udev rule is to just "touch" it in /etc/udev/rule.d/ (i. e. empty file), and if
the rule is customized we cannot really predict anyway if the user wants
MAC-based USB net names or not.

LP: #1615021