fsckd: Do not exit on idle timeout if there are still clients connected
systemd-fsckd's event loop terminates if nothing happens for 30 seconds.
Exiting prematurely while fsck is still running but simply too slow to
send us progress updates would close the socket and fsck would receive
SIGPIPE when it writes to the socket. If this happens, the fsck process
is aborted and the file system check is not completed.

Closes: #788050
LP: #1547844