Backport networkd 231
Compared to 229 this has a lot of fixes, some of which we need for good netplan
support. Backporting them individually would be a lot more work and a lot less
robust, and we did not use/support networkd in 16.04 so far.

Drop the other network related patches as they are included in this backport
now.

(LP: #1627641)