Add Conflicts against hal
Since v183, udev no longer supports RUN+="socket:". This feature is
still used by hal, but now generates vast amounts of errors in the
journal. Thus force the removal of hal by adding a Conflicts to the udev
package. This is safe, as hal is long dead and no package depends on it
anymore.