init-functions Only call daemon-reload when planning to redirect
systemctl daemon-reload is a quite a heavy operation, it will re-parse
all configuration and re-run all generators. This should only be done
when strictly needed.

The init-function helpers try to cope with an lsb init script being
called *before* the sysv generator has run. Unfortunately that always got
triggered when other script sourced the helpers as there won't be a
service matching them.

Resolve this by only potentially reloading the daemon when preparing to
redirect.

Closes: #861158