Exclude test binaries from dh_shlibdeps
The test binaries in libsystemd-dev require libsystemd-shared which is
shipped in the systemd package. Those test binaries are primarily meant
to be run via autopkgtest. As the libsystemd-dev package is not supposed
to depend on systemd, exclude the tests from dh_shlibdeps and instead
update the autopkgtest dependencies to pull in the systemd package.

Closes: #859152