Enable BPF support
Use clang for now, as gcc-bpf needs some bug fixes to work.

Note that Ubuntu's packaging of bpftool is difficult to use, so
first of all we will always need to leave this in d/control even
after moving to a release that needs it:

 linux-tools-generic <pkg.systemd.upstream> | bpftool <pkg.systemd.upstream>,

That is because bpftool is not enough on Ubuntu, linux-tools-generic
is needed, but we don't have it in Debian. But we can't remove the
build profile, as on the buildd this will always fail. We need it
for upstream builds.

Gbp-Dch: ignore