Drop out-of-tree coredump patch and use config drop-ins instead
As per: https://github.com/systemd/systemd/pull/27066
instead of carrying an out-of-tree patch, use drop-ins to set the
default pattern to 'core' and the default core limit to '0' in
the systemd package, and a higher priority drop-in to set the
default core limit to 'infinity' in the systemd-coredump package
(it already ships a sysctl to set the pattern). The end result is
the same, but with config drop-ins instead of patches.